Fixed search cron job to remove slashes

+++ public/cron/words.php
@@ -15,59 +15,51 @@ if(mysqli_num_rows($result))
 
     $page = mysqli_fetch_assoc($result);
 
-    echo '<h1>Scanning: '.$page['url'].'</h1>';
-    
-    $status = url_status($page['url']);
-
-    echo '<h2>Error Code: ',$status.'</h2>';
-
-    $query = 'DELETE FROM page_word
-        WHERE page_id = "'.$page['id'].'"';
-    mysqli_query($connect, $query);
-
     $query = 'UPDATE pages SET
-        status = "'.$status.'",
         updated_at = NOW(),
         scrapped_at = NOW()
         WHERE id = '.$page['id'].'
         LIMIT 1';
-    mysqli_query($connect, $query);
+    mysqli_query($connect, $query); 
 
-    if($status && $status == '200')
+    echo '<h1>Scanning: '.$page['url'].'</h1>';
+    
+    $status = url_status($page['url']);
+
+    if($status)
     {
 
-        $html = url_content($page['url']);
-        $title = html_fetch_title($html);
-        $words = html_fetch_words($html);
+        echo '<h2>Error Code: ',$status.'</h2>';
+
+        $query = 'DELETE FROM page_word
+            WHERE page_id = "'.$page['id'].'"';
+        mysqli_query($connect, $query);
 
         $query = 'UPDATE pages SET
-            title = "'.$title.'"
+            status = "'.$status.'"
             WHERE id = '.$page['id'].'
             LIMIT 1';
         mysqli_query($connect, $query);
 
-        echo '<h2>Words:</h2>';
-
-        foreach($words as $word => $count)
+        if($status && $status == '200')
         {
 
-            $word = mysqli_real_escape_string($connect, $word);
+            $html = url_content($page['url']);
+            $title = html_fetch_title($html);
+            $words = html_fetch_words($html);
 
-            $query = 'SELECT *
-                FROM words
-                WHERE word = "'.$word.'"
+            $query = 'UPDATE pages SET
+                title = "'.$title.'"
+                WHERE id = '.$page['id'].'
                 LIMIT 1';
-            $result = mysqli_query($connect, $query);
+            mysqli_query($connect, $query);
 
-            if(!mysqli_num_rows($result))
+            echo '<h2>Words:</h2>';
+
+            foreach($words as $word => $count)
             {
 
-                $query = 'INSERT INTO words (
-                        word
-                    ) VALUES (
-                        "'.$word.'"
-                    )';
-                mysqli_query($connect, $query);
+                $word = mysqli_real_escape_string($connect, $word);
 
                 $query = 'SELECT *
                     FROM words
@@ -75,26 +67,44 @@ if(mysqli_num_rows($result))
                     LIMIT 1';
                 $result = mysqli_query($connect, $query);
 
+                if(!mysqli_num_rows($result))
+                {
+
+                    $query = 'INSERT INTO words (
+                            word
+                        ) VALUES (
+                            "'.$word.'"
+                        )';
+                    mysqli_query($connect, $query);
+
+                    $query = 'SELECT *
+                        FROM words
+                        WHERE word = "'.$word.'"
+                        LIMIT 1';
+                    $result = mysqli_query($connect, $query);
+
+                }
+
+                $word = mysqli_fetch_assoc($result);
+
+                $query = 'INSERT INTO page_word (
+                        word_id,
+                        page_id,
+                        count
+                    ) VALUES (
+                        "'.$word['id'].'",
+                        "'.$page['id'].'",
+                        "'.$count.'"
+                    )';
+                mysqli_query($connect, $query);
+
+                echo '<hr>';
+                echo $query;
+
             }
-
-            $word = mysqli_fetch_assoc($result);
-
-            $query = 'INSERT INTO page_word (
-                    word_id,
-                    page_id,
-                    count
-                ) VALUES (
-                    "'.$word['id'].'",
-                    "'.$page['id'].'",
-                    "'.$count.'"
-                )';
-            mysqli_query($connect, $query);
-
-            echo '<hr>';
-            echo $query;
-
+            
         }
-        
+
     }
     
 }
